# backend/docs/swagger.yaml
basePath: /
definitions:
...
  models.Permission:
    properties:
      button_id:
        example: 1
        type: integer
      created_at:
        type: string
      id:
        example: 1
        type: integer
      menu_id:
        example: 1
        type: integer
      name:
        example: 查看仪表盘
        type: string
      permission_code:
        example: view_dashboard
        type: string
      roles:
        items:
          $ref: '#/definitions/models.Role'
        type: array
      type:
        example: menu
        type: string
      updated_at:
        type: string
    type: object
  models.Role:
    properties:
      created_at:
        type: string
      description:
        example: 系统管理员角色
        type: string
      id:
        example: 1
        type: integer
      name:
        example: 管理员
        type: string
      permissions:
        items:
          $ref: '#/definitions/models.Permission'
        type: array
      updated_at:
        type: string
      users:
        items:
          $ref: '#/definitions/models.User'
        type: array
    type: object
  models.User:
    properties:
      created_at:
        type: string
      id:
        example: 1
        type: integer
      password:
        example: password123
        type: string
      roles:
        items:
          $ref: '#/definitions/models.Role'
        type: array
      updated_at:
        type: string
      username:
        example: admin
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: 租户中心系统的API文档
  title: 租户中心API
  version: "1.0"
paths:
...
  /api/permissions:
    get:
      consumes:
      - application/json
      description: 获取所有权限的列表，需要管理员权限
      produces:
      - application/json
      responses:
        "200":
          description: 权限列表
          schema:
            items:
              $ref: '#/definitions/models.Permission'
            type: array
        "500":
          description: 获取权限列表失败
          schema:
            type: object
      security:
      - ApiKeyAuth: []
      tags:
      - 权限管理
    post:
      consumes:
      - application/json
      description: 创建新权限，需要管理员权限
      parameters:
      - description: 权限信息
        in: body
        name: permission
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: 权限创建成功
          schema:
            type: object
        "400":
          description: 无效的请求参数
          schema:
            type: object
        "500":
          description: 创建权限失败
          schema:
            type: object
      security:
      - ApiKeyAuth: []
      tags:
      - 权限管理
  /api/permissions/{id}:
    put:
      consumes:
      - application/json
      description: 更新指定权限的信息，需要管理员权限
      parameters:
      - description: 权限ID
        in: path
        name: id
        required: true
        type: integer
      - description: 权限信息
        in: body
        name: permission
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 权限信息更新成功
          schema:
            type: object
        "400":
          description: 无效的请求参数
          schema:
            type: object
        "500":
          description: 更新权限信息失败
          schema:
            type: object
      security:
      - ApiKeyAuth: []
      tags:
      - 权限管理
  /api/permissions/type/{type}:
    get:
      consumes:
      - application/json
      description: 获取指定类型的所有权限列表，需要管理员权限
      parameters:
      - description: 权限类型
        in: path
        name: type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 权限列表
          schema:
            items:
              $ref: '#/definitions/models.Permission'
            type: array
        "500":
          description: 获取权限列表失败
          schema:
            type: object
      security:
      - ApiKeyAuth: []
      tags:
      - 权限管理
  /api/roles:
    get:
      consumes:
      - application/json
      description: 获取所有角色的列表，需要管理员权限
      produces:
      - application/json
      responses:
        "200":
          description: 角色列表
          schema:
            items:
              $ref: '#/definitions/models.Role'
            type: array
        "500":
          description: 获取角色列表失败
          schema:
            type: object
      security:
      - ApiKeyAuth: []
      tags:
      - 角色管理
    post:
      consumes:
      - application/json
      description: 创建新角色，需要管理员权限
      parameters:
      - description: 角色信息
        in: body
        name: role
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: 角色创建成功
          schema:
            type: object
        "400":
          description: 无效的请求参数
          schema:
            type: object
        "500":
          description: 创建角色失败
          schema:
            type: object
      security:
      - ApiKeyAuth: []
      tags:
      - 角色管理
  /api/roles/{id}:
    put:
      consumes:
      - application/json
      description: 更新指定角色的信息，需要管理员权限
      parameters:
      - description: 角色ID
        in: path
        name: id
        required: true
        type: integer
      - description: 角色信息
        in: body
        name: role
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 角色信息更新成功
          schema:
            type: object
        "400":
          description: 无效的请求参数
          schema:
            type: object
        "500":
          description: 更新角色信息失败
          schema:
            type: object
      security:
      - ApiKeyAuth: []
      tags:
      - 角色管理
  /api/roles/{id}/permissions:
    get:
      consumes:
      - application/json
      description: 获取指定角色的所有权限列表，需要管理员权限
      parameters:
      - description: 角色ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 权限列表
          schema:
            items:
              $ref: '#/definitions/models.Permission'
            type: array
        "400":
          description: 无效的角色ID
          schema:
            type: object
        "500":
          description: 获取角色权限列表失败
          schema:
            type: object
      security:
      - ApiKeyAuth: []
      tags:
      - 角色管理
    post:
      consumes:
      - application/json
      description: 为指定角色绑定一个或多个权限，需要管理员权限
      parameters:
      - description: 角色ID
        in: path
        name: id
        required: true
        type: integer
      - description: 权限ID列表
        in: body
        name: permissionIDs
        required: true
        schema:
          items:
            type: integer
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: 权限绑定成功
          schema:
            type: object
        "400":
          description: 无效的请求参数
          schema:
            type: object
        "500":
          description: 绑定权限失败
          schema:
            type: object
      security:
      - ApiKeyAuth: []
      tags:
      - 角色管理
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"